Use random.nextLong() when generating a new ErrorHeader in MessageBuilder. Fix typo epadEmail -> epadPassword in EmailPasswordAuthenticationFactory. Fix some JUnit errors assertEquals() -> assertArrayEquals(). Minor FindBugs fixes.

+++ src/main/java/com/netflix/msl/msg/MessageBuilder.java
@@ -388,7 +388,7 @@ public class MessageBuilder {
         else {
             final Random random = ctx.getRandom();
             do {
-                messageId = random.nextInt();
+                messageId = random.nextLong();
             } while (messageId < 0 || messageId > MslConstants.MAX_LONG_VALUE);
         }
         final ResponseCode errorCode = error.getResponseCode();
